Miscellaneous tinkering
I made the exit happen on case 7 instead of 0, because exiting the program wasn't working for me. As much as I could with the power being out, I also tested data storage and retrieval some more, and I've started working on making findCustomerById work once again.
Right now I'm not sure how I will, though, because I need to create a customer from the process of reading the rental records, but I don't have an ID to recreate the customer with.

Sorry this is such a dense commit with so many different things, I've been trying to avoid that, it's just that I couldn't commit anything at all for 2 days.