Fixed the issues with findCustomerById
I have everything in regards to finding customers by their Ids working now, and I've refactored the code for adding rental records to now loop through all customers in the list, compare their names to the ones in the rental record being looked at, and then accessing the ID of the customer that way. It's still not accurately storing the rental history when asked from what I can tell, however.