Prevented repeat vehicles and customers from being added to the system
Essentially, I did this by just looking at the license plates/IDs of the vehicles/customers being added, comparing them to each of the ones already in the list. If one is found to be the same, we treat it as the same vehicle/customer as one that already exists, and then do not add them to our lists, or save them to our ser files.