Refactored capitalization logic in the Vehicle constructor
The new method, capitalize, will do the same thing as the code above- checking if a string is null or empty, then capitalizing the first character while making all others lowercase if not- to any string we plug in. Calling this method on the make and model when we make a vehicle should improve clarity while accomplishing the same thing.