Rent and return testing
I added new test cases for renting and returning vehicles, to make sure it's working properly. I then had to clear the .ser files, because I wanted to run both that test case and the license plate validation test. Both went off without a hitch. The only thing is that the .ser file didn't recreate itself- for one reason or another, the test customers don't seem to get saved to the .ser file. It's of no serious concern to me, just bringing it up for transparency.